CNAM-54 Created Comparator to check flatenning results
CNAM-54 Address review comments.

CNAM-54 corrected === function in RichDataFrames.

CNAM-54 updated comparatorSuite according to review comments.

CNAM-54 Problem with expected results of runETL test in MainSuite.

CNAM-54 Corrected test for runETL in MainSuite

CNAM-54 Debugging travis

CNAM-54 Corrected MainSuite expected values

+++ src/test/scala/fr/polytechnique/cmap/cnam/filtering/DrugEventsTransformerSuite.scala
@@ -6,6 +6,7 @@ import fr.polytechnique.cmap.cnam.SharedContext
 import fr.polytechnique.cmap.cnam.utilities.RichDataFrames
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.functions._
+import org.apache.spark.sql.types.IntegerType
 
 
 class DrugEventsTransformerSuite extends SharedContext {
@@ -33,7 +34,7 @@ class DrugEventsTransformerSuite extends SharedContext {
       .select(
         col("PHA_PRS_IDE").as("CIP07"),
         col("MOLECULE_NAME"),
-        col("TOTAL_MG_PER_UNIT")
+        col("TOTAL_MG_PER_UNIT").cast(IntegerType)
       )
     val expected = Seq(
       ("patient", "3541848", "3400935418487", "GLICLAZIDE", 900),
